feat: split god file into smaller files
